example: update generated files
